Fix accessibility aria-labels and extract shared linkPersonDataById helper
- Add aria-label='Add spouse' to spouse icon button in person-family-list
- Add aria-label='Add child' to child icon buttons in both components
- Extract duplicated linkPersonDataById into LinkPersonDialogData.fromPersonId() static factory
- Remove duplicated private methods and unused LinkPersonItem imports
