+++ gedbrowserng-frontend/src/app/models/link-person-dialog-data.ts
@@ -10,6 +10,20 @@ export class LinkPersonDialogData {
   selectOne: LinkPersonItem;
   dataset: string;
   titleString: string;
+
+  static fromPersonId(personId: string): LinkPersonDialogData {
+    const selectedPerson = new ApiPerson();
+    selectedPerson.string = personId;
+    const selected: LinkPersonItem = {
+      id: personId,
+      label: personId,
+      person: selectedPerson
+    };
+    const data = new LinkPersonDialogData();
+    data.selected = [selected];
+    data.selectOne = selected;
+    return data;
+  }
 }
 
 export class LinkPersonItem {
